requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
python-pam==2.0.2
gunicorn==21.2.0
Werkzeug==3.0.1
